:recycle: refactor(review): split review write form script
